LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-tokenization-top-left-expected.txt


FAIL "left=141" should set left position of opened window assert_equals: expected 141 but got 152
FAIL " left = 141" should set left position of opened window assert_equals: expected 141 but got 152
FAIL "left==141" should set left position of opened window assert_equals: expected 141 but got 152
FAIL "
left= 141" should set left position of opened window assert_equals: expected 141 but got 152
FAIL ",left=141,," should set left position of opened window assert_equals: expected 141 but got 152
FAIL "LEFT=141" should set left position of opened window assert_equals: expected 141 but got 152
FAIL "top=142" should set top position of opened window assert_equals: expected 142 but got 152
FAIL " top = 142" should set top position of opened window assert_equals: expected 142 but got 152
FAIL "top==142" should set top position of opened window assert_equals: expected 142 but got 152
FAIL "\ttop= 142" should set top position of opened window assert_equals: expected 142 but got 152
FAIL ",top=142,," should set top position of opened window assert_equals: expected 142 but got 152
FAIL "TOP=142" should set top position of opened window assert_equals: expected 142 but got 152
PASS "top=152,left=152" should set top and left position of opened window
PASS "top=152,,left=152," should set top and left position of opened window
PASS "top=152==left=152" should set top and left position of opened window
PASS ",,top= 152, left=152" should set top and left position of opened window
